Fixed sign comparison warning

+++ Source/DKUnicode.c
@@ -192,7 +192,7 @@ size_t dk_ustrnlen( const char * str, size_t n )
     size_t length = 0;
     const char * cur = str;
 
-    while( (*cur != '\0') && ((cur - str) < n) )
+    while( (*cur != '\0') && ((size_t)(cur - str) < n) )
     {
         DKChar32 ch;
         cur += dk_ustrscan( cur, &ch );
